Add files via upload
 Gross Domestic Product (GDP) analysis using data science is a critical and multifaceted 
approach to understanding and evaluating the economic performance of a country or region. GDP serves as a 
fundamental metric for assessing the overall health and growth of an economy, and data science techniques have 
become indispensable in extracting valuable insights from the vast amount of data generated in modern 
economies. The abstract provides an overview of how data science is applied to GDP analysis, highlighting its 
significance, methodologies, and key takeaways.GDP is a comprehensive measure of a nation's economic 
activity, encompassing the total value of goods and services produced within a specific time frame. It serves as a 
crucial indicator for policymakers, businesses, and investors to make informed decisions. conclusion, GDP 
analysis with data science represents a powerful approach to understanding and managing economic 
performance. By harnessing the capabilities of data science, stakeholders can gain deeper insights, enhance 
decision-making, and adapt more effectively to the dynamic nature of modern economies